requirements.txt
runpod>=1.0.0
torch>=2.8.0
torchvision>=0.23.0
torchaudio>=2.8.0
diffusers>=0.35.2
transformers>=4.57.1
accelerate>=1.11.0
safetensors>=0.6.2
pillow>=10.0.0
numpy>=1.24.0
trimesh>=4.0.0
rembg>=2.0.0
huggingface-hub>=0.36.0
pymeshlab>=2023.12
opencv-python>=4.8.0
scipy>=1.11.0
scikit-image>=0.21.0
tqdm>=4.66.0
requests>=2.32.0
